requirements.txt
requests
beautifulsoup4
lxml
selenium
webdriver-manager
openpyxl
plotly
supabase
python-dotenv